requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.4
django-crispy-forms==2.3
django-filter==24.3
html5lib==1.1
idna==3.10
lxml==5.3.0
oscrypto==1.3.0
pillow==11.0.0
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
python-bidi==0.6.3
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
svglib==1.5.1
tinycss2==1.4.0
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
